+++ GA-AppLocker/Modules/GA-AppLocker.Setup/GA-AppLocker.Setup.psm1
@@ -152,6 +152,7 @@ Export-ModuleMember -Function @(
     'Initialize-AppLockerEnvironment',
     'Get-SetupStatus',
     'Enable-WinRMGPO',
-    'Disable-WinRMGPO'
+    'Disable-WinRMGPO',
+    'Remove-WinRMGPO'
 )
 #endregion